refactor: move intersection of geometry curves to single file
